Updated MockupDataGenerator to V2.
