+++ app/src/main/java/com/koleff/kare_android/ui/compose/components/SearchExercisesList.kt
@@ -10,7 +10,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.koleff.kare_android.common.MockupDataGenerator
+import com.koleff.kare_android.common.MockupDataGeneratorV2
 import com.koleff.kare_android.data.model.dto.ExerciseDto
 import com.koleff.kare_android.ui.compose.banners.ExerciseBannerV2
 import com.koleff.kare_android.ui.compose.banners.MuscleGroupHeader
@@ -66,7 +66,7 @@ fun SearchExercisesListPreview() {
         .fillMaxSize()
         .padding(8.dp)
 
-    val exerciseList = MockupDataGenerator.generateExerciseList()
+    val exerciseList = MockupDataGeneratorV2.generateExerciseList()
 
     SearchExercisesList(
         modifier = modifier,
